requirements.txt
streamlit
pandas
faiss-cpu
sentence-transformers 
jax==0.4.14
jaxlib==0.4.14